% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/known_age.R
\name{age.comp.plot}
\alias{age.comp.plot}
\title{Plots of quality measures of age estimation}
\usage{
age.comp.plot(
  x,
  age_identifier = "age.s",
  known_age,
  mean_choice = "Mode",
  hdi_color = c("chartreuse4", "coral2")
)
}
\arguments{
\item{x}{output from the function \code{diagnostic.summary()}.}

\item{age_identifier}{a character string of either "age.s" or "age.s_c" to
select the uncalibrated or calibrated age estimates. Default: "age.s".}

\item{known_age}{a vector of known age-at-death. NAs are allowed and those
entries will subsequently be ignored.}

\item{mean_choice}{a character string of either "Mean", "Median" or "Mode".
Default: "Mode".}

\item{hdi_color}{a character vector of exactly two entries with color values
to differentiate estimated ages within the HDI from those outside the HDI.
Default: c("chartreuse4", "coral2")}
}
\value{
A ggplot object with 2 x 2 single plots, showing:
\describe{
  \item{Top left}{Comparison of estimated \emph{highest density intervals}
  with known ages, color1 = age within HDI, color2 = age outside HDI,
  individuals ordered according to known age-at-death. }
  \item{Top right}{Comparison of the density of known ages with a Gompertz
  function derived from the arithmetic mean of the estimated population
  parameters \eqn{\alpha} and \eqn{\beta}.}
  \item{Bottom left}{Scatter plot of known and estimated ages with
  regression line in blue. The dotted line marks perfect equivalence.}
  \item{Bottom right}{Slope of the regression line from the left bottom image
  (cf. goodness-of-fit measure \code{Residual_slope} from the function
  \code{age.comp.summary()}).}
}
}
\description{
Visualisation of the difference between estimated and known age with the
help of a combination of plots.
}
\examples{
\dontshow{if (interactive()) withAutoprint(\{ # examplesIf}

  # select Spitalfields data with multiple traits
  spitalfields_traits <- spitalfields[,c(2:6)]

  # example with multinormal likelihood, please be patient
  spitalfields_res <- bay.ta(algorithm = "mnorm",
  method = spitalfields_traits)

  # compute age summary statistics
  age.comp.plot(spitalfields_res, known_age = spitalfields$Age)
\dontshow{\}) # examplesIf}
}
