% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/known_age.R
\name{age.comp.summarize}
\alias{age.comp.summarize}
\title{Quality measures of age estimation}
\usage{
age.comp.summarize(
  x,
  selector = NULL,
  mean_choice = "Mode",
  known_age,
  age_identifier = "age.s",
  ...
)
}
\arguments{
\item{x}{MCMC output from coda chains.}

\item{selector}{a vector of individuals to include. Optional.}

\item{age_identifier}{a character string of either "age.s" or "age.s_c" to
select the uncalibrated or calibrated age estimates. Default: "age.s".}

\item{chosen_mean}{a character string of either "Mode", "Median" or "Mode".
Default: "Mode".}
}
\value{
a dataframe with age estimation quality parameters
}
\description{
Comparison of estimated age with known age-at-death.
}
\examples{
NULL

}
