% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/known_age.R
\name{age.comp.summary}
\alias{age.comp.summary}
\title{Quality measures of age estimation}
\usage{
age.comp.summary(
  mcmc_list,
  known_age,
  mean_choice = "Mode",
  age_identifier = "age.s",
  ...
)
}
\arguments{
\item{mcmc_list}{MCMC output from coda chains.}

\item{known_age}{a vector of known age-at-death. NAs are allowed and those
entries will subsequently be ignored.}

\item{mean_choice}{a character string of either "Mode", "Median" or "Mode".
Default: "Mode".}

\item{age_identifier}{a character string of either "age.s" or "age.s_c" to
select the uncalibrated or calibrated age estimates. Default: "age.s".}

\item{...}{
  Arguments passed on to \code{\link[=diagnostic.summary]{diagnostic.summary}}
  \describe{
    \item{\code{HDImass}}{numeric. Value within 0 and 1. Default = 0.95.}
    \item{\code{gelman_diag}}{logical. If TRUE, the Gelman-Rubin diagnostics for computing
the PSRF is invoked. Default: TRUE.}
  }}
}
\value{
a dataframe with age estimation quality parameters
}
\description{
Comparison of estimated age with known age-at-death.
}
\examples{
NULL

}
