% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayes_helper.R
\name{age.estim.summary}
\alias{age.estim.summary}
\title{Summary of age estimates}
\usage{
age.estim.summary(x, mean_choice = "Mode", age_identifier = "age.s")
}
\arguments{
\item{x}{output from function `diagnostics.summary()`}

\item{mean_choice}{a character string of either "Mode", "Median" or "Mode".
Default: "Mode".}

\item{age_identifier}{a character string of either "age.s" or "age.s_c" to
select the uncalibrated or calibrated age estimates. Default: "age.s".}
}
\value{
a vector with M, b, a, HDI, quantile HDI range
}
\description{
Median ranges
}
\examples{
NULL

}
