% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/known_age.R
\name{age.estimate.summarize}
\alias{age.estimate.summarize}
\title{Quality measures of age estimation}
\usage{
age.estimate.summarize(
  x,
  selector = NA,
  chosen_mean = "Mode",
  known_age,
  age_identifier = "age.s",
  HDImass = 0.95,
  gelman_diag = TRUE,
  gelman_diag_multivariate = F
)
}
\arguments{
\item{x}{MCMC output from coda chains.}

\item{selector}{a vector of individuals to include. Optional.}

\item{chosen_mean}{a character string of either "Mode", "Median" or "Mode".
Default: "Mode".}

\item{age_identifier}{a character string of either "age.s" or "age.s_c" to
select the uncalibrated or calibrated age estimates. Default: "age.s".}
}
\value{
a dataframe with age estimation quality parameters
}
\description{
Comparison with known age-at-death
}
\examples{
NULL

}
