% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bay_ta_jags.R
\name{bay.ta.jags}
\alias{bay.ta.jags}
\title{Bayesian Transition Analysis with JAGS}
\usage{
bay.ta.jags(
  method,
  parameters,
  gomp_b = NA,
  minimum_age = 15,
  maximum_age = 100,
  error_sd = NA,
  adaptSteps = 2000,
  burnInSteps = 3000,
  runjagsMethod = "rjags",
  nChains = 3,
  thinSteps = 1,
  numSavedSteps = 10000,
  silent.jags = F,
  silent.runjags = F,
  seed = seed
)
}
\arguments{
\item{method}{matrix of integers, converted to matrix if not already matrix.
Ordinal trait(s) for age estimation.}

\item{parameters}{vector of character strings. Parameters to monitor.}

\item{gomp_b}{numeric. Optional prior for parameter Gompertz beta. Default:
\code{NA}.}

\item{minimum_age}{numeric. Minimum age for Gompertz distribution. Default:
\code{15}.}

\item{maximum_age}{numeric. Maximum age for Gompertz distribution. Default:
\code{100}.}

\item{error_sd}{numeric. Optional error parameter for age estimates. Default:
\code{NA}.}

\item{adaptSteps}{integer. Number of adaptation steps, ignored when
\code{framework} is set to \code{NIMBLE}. Default: \code{2000}.}

\item{burnInSteps}{integer. Number of steps for burn-in. Default: \code{3000}.}

\item{runjagsMethod}{string. Mode to run `runjags`, options: "rjags",
"rjparallel", "parallel". Default: "rjags".}

\item{nChains}{integer. Number of chains. Default: \code{3}.}

\item{thinSteps}{integer. Thinning, i. e. which \emph{i}th step should be
saved. Default: \code{1} (no thinning).}

\item{numSavedSteps}{integer. Number of saved steps. Default: \code{10000}.
The total number of steps equals \code{thinSteps × numSavedSteps}.}

\item{silent.jags}{TRUE/FALSE Silent mode to run JAGS. Default: \code{FALSE}.
Ignored when \code{framework} is set to \code{NIMBLE}.}

\item{silent.runjags}{TRUE/FALSE Silent mode to run runjags. Default:
\code{FALSE}. Ignored when \code{framework} is set to \code{NIMBLE}.}

\item{seed}{integer. Random number for reproducibility. In parallel
processing, each cluster automatically gets different seeds. If no seed is
specified, the value is set to today's date as integer.}
}
\value{
A list of MCMC chains of class \code{coda::mcmc.list}.
}
\description{
Bayesian Transition Analysis with JAGS
}
\examples{
\dontshow{if (interactive()) withAutoprint(\{ # examplesIf}

  # select Sorsum data with auricular surface after Lovejoy et al. 1985 and
  # convert to matrix
  sorsum <- as.matrix(sorsum_as[,2])

  # example with framework JAGS
  sorsum_res <- bay.ta(framework = "JAGS", method = sorsum)
\dontshow{\}) # examplesIf}
}
\keyword{internal}
