% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bay_ta_nimble_mnorm.R
\name{bay.ta.nimble.mnorm}
\alias{bay.ta.nimble.mnorm}
\title{Multinormal bayta}
\usage{
bay.ta.nimble.mnorm(
  method,
  parameters,
  eta = 1,
  gomp_b = NA,
  minimum_age = 15,
  maximum_age = 100,
  burnInSteps = 2000,
  nChains = 3,
  thinSteps = 1,
  numSteps = 10000,
  seed = FALSE
)
}
\arguments{
\item{method}{matrix.}

\item{parameters}{parameters}

\item{eta}{Cholesky}

\item{gomp_b}{Gompertz}

\item{minimum_age}{minimum age}

\item{maximum_age}{maximum age}

\item{burnInSteps}{burn-in}

\item{nChains}{number of chains}

\item{thinSteps}{thinning}

\item{numSteps}{number of steps}

\item{seed}{seed}
}
\value{
coda object
}
\description{
Multinormal bayta
}
\examples{
NULL

}
