% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bay_ta_nimble_norm.R
\name{bay.ta.nimble.norm}
\alias{bay.ta.nimble.norm}
\title{Nimble bayta}
\usage{
bay.ta.nimble.norm(
  method,
  parameters,
  gomp_b = NA,
  minimum_age = 15,
  maximum_age = 100,
  burnInSteps = 2000,
  nChains = 3,
  thinSteps = 1,
  numSteps = 10000,
  seed = FALSE
)
}
\arguments{
\item{method}{matrix}

\item{parameters}{parameters}

\item{gomp_b}{Gompertz}

\item{minimum_age}{min age}

\item{maximum_age}{maxage}

\item{burnInSteps}{burn-in}

\item{nChains}{number of hains}

\item{thinSteps}{thinning}

\item{numSteps}{number of steps}

\item{seed}{seed}
}
\value{
coda object
}
\description{
Nimble bayta
}
\examples{
NULL

}
