% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/known_age.R
\name{bay.ta.plot}
\alias{bay.ta.plot}
\title{Plots of quality measures of age estimation}
\usage{
bay.ta.plot(x, selector = NA, age_identifier = "age.s", known_age, mean_choice)
}
\arguments{
\item{x}{output from the function `diagnostic.summary`}

\item{selector}{a vector of individuals to include. Optional.}

\item{age_identifier}{a character string of either "age.s" or "age.s_c" to
select the uncalibrated or calibrated age estimates. Default: "age.s".}

\item{mean_choice}{a character string of either "Mode", "Median" or "Mode".
Default: "Mode".}
}
\value{
a ggplot object with 2 x 2 single plots, showing
}
\description{
Comparison with known age-at-death
}
\examples{
NULL

}
