% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bay_ta.R
\name{bay.ta}
\alias{bay.ta}
\title{Bayesian Transition Analysis}
\usage{
bay.ta(
  framework = "JAGS",
  algorithm = "norm",
  multicore = FALSE,
  seed = as.integer(format(Sys.Date(), "\%Y\%m\%d")),
  method,
  eta = 1,
  gomp_b = NA,
  error_sd = NA,
  minimum_age = 15,
  maximum_age = 100,
  parameters = c("b", "a", "beta0", "beta", "thresh", "age.s"),
  nChains = 3L,
  adaptSteps = 2000,
  burnInSteps = 3000,
  thinSteps = 1,
  numSavedSteps = 10000,
  silent.jags = F,
  silent.runjags = F
)
}
\arguments{
\item{algorithm}{character string. Either \code{norm} for 'simple' ordered
regression or \code{mnorm} for multinormal ordered regression. Default:
\code{norm}.}

\item{multicore}{TRUE/FALSE. If TRUE each chain is assigned to a dedicated
core. Default: FALSE.}

\item{seed}{integer. Random number for reproducibility. In parallel
processing, each cluster automatically gets different seeds. If no seed is
specified, the value is set to today's date as integer.}

\item{method}{matrix of integers. Ordinal trait(s) for age estimation.}

\item{eta}{Cholesky}

\item{gomp_b}{double. Optional prior for parameter Gompertz beta. Default:
NA.}

\item{minimum_age}{double. Minimum age for Gompertz distribution. Default:
15.}

\item{maximum_age}{double. Maximum age for Gompertz distribution. Default:
100.}

\item{parameters}{vector of character strings. Parameters to monitor.}

\item{nChains}{integer. Number of chains. Default: 3.}

\item{adaptSteps}{integer. Number of adaptation steps, ignored when
framework is set to NIMBLE. Default: 2000.}

\item{burnInSteps}{integer. Number of steps for burn-in. Default: 3000.}

\item{thinSteps}{integer. Thinning, i. e. which ith step should be saved.
Default: 1 (no thinning).}

\item{numSavedSteps}{integer. Number of saved steps. Default: 10000. The
total number of steps equals \code{thinSteps × numSavedSteps}.}

\item{silent.jags}{TRUE/FALSE Silent mode to run JAGS. Default: FALSE.
Ignored when framework is set to NIMBLE.}

\item{silent.runjags}{TRUE/FALSE Silent mode to run runjags. Default: FALSE.
Ignored when framework is set to NIMBLE.}
}
\value{
A list of MCMC chains of class `coda::mcmc.list`.
}
\description{
This is a wrapper for the functions \code{bay.ta.jags} and
\code{bay.ta.nimble}. Both implement the Bayesian Transition Analysis
with MCMC. The framework NIMBLE allows the user to run models with
multinormal ordered regression, also with parallel clusters while JAGS
tends to be more stable. The latter presupposes, however, that you have
installed JAGS outside of R.
}
\details{
blabla
}
\examples{
# select Sorsum data with auricular surface after Lovejoy et al. 1985 and
# convert to matrix
sorsum <- as.matrix(sorsum_as[,2])

# example with default settings
\dontrun{
sorsum_res <- bay.ta(method = sorsum) }

# example with framework NIMBLE
\dontrun{
sorsum_res <- bay.ta(framework = "NIMBLE", method = sorsum) }

# example with framework JAGS and multiple cores (parallel computing)
\dontrun{
sorsum_res <- bay.ta(framework = "JAGS", multicore = TRUE, method = sorsum) }

# example with 100,000 saved iterations and a thinning of 100
\dontrun{
sorsum_res <- bay.ta(method = sorsum, numSavedSteps = 100000, thin = 100) }

# select Spitalfields data with multiple traits and convert to matrix
spitalfields_traits <- as.matrix(spitalfields[,c(2:6)])

# example with multinormal likelihood
\dontrun{
spitalfields_res <- bay.ta(framework = "NIMBLE", algorithm = "mnorm",
method = spitalfields_traits) }

}
