% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayes_helper.R
\name{corr.mat.mean}
\alias{corr.mat.mean}
\title{Extract correlation matrix from Cholesky factor}
\usage{
corr.mat.mean(mcmc_list)
}
\arguments{
\item{mcmc_list}{MCMC output from coda chains.}
}
\value{
a matrix with correlations between traits
}
\description{
As the LKJ prior uses the Cholesky decomposition of the correlation matrix,
getting the correlation indices from the coda chains is less straightforward
than it seems. It involves taking the cross product from the resulting coda
estimates.
}
\examples{
NULL

}
