% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gompertz_nimble.R
\name{dgomp}
\alias{dgomp}
\alias{rgomp}
\alias{pgomp}
\alias{qgomp}
\title{Gompertz probability distribution}
\usage{
rgomp(n, b, a)

dgomp(x, b, a, log = 0)

pgomp(q, b, a, lower.tail, log.p = 0)

qgomp(p, b, a, lower.tail, log.p = 0)
}
\arguments{
\item{n}{n}

\item{b}{b}

\item{a}{a}

\item{x}{x}

\item{log}{log}

\item{q}{q}

\item{lower.tail}{lower.tail}

\item{log.p}{log.p}

\item{p}{p}
}
\value{
depending on the kind of function, either
}
\description{
The Gompertz distribution and Gompertz function for internal use in
\code{NIMBLE} models
}
