% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayes_helper.R
\name{diagnostics.max.min}
\alias{diagnostics.max.min}
\title{Maximum and minimum diagnostic values}
\usage{
diagnostics.max.min(x)
}
\arguments{
\item{x}{output from function `diagnostics.summary()`}
}
\description{
max. and min diagnostics values
}
\examples{
NULL

}
