% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayes_helper.R
\name{extract.corr}
\alias{extract.corr}
\title{Extract correlation matrix from Cholesky factor}
\usage{
extract.corr(x)
}
\arguments{
\item{x}{matrix. Output from coda chains}
}
\value{
a matrix with correlations between traits
}
\description{
Function for computation of the correlation matrix from Cholesky factor
}
\examples{
NULL

}
