% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayes_helper.R
\name{extract_corr_from_cholesky}
\alias{extract_corr_from_cholesky}
\title{Extract correlation matrix from Cholesky factor}
\usage{
extract_corr_from_cholesky(x)
}
\arguments{
\item{x}{matrix. Output from coda chains}

\item{n_methods}{integer. Number of traits}
}
\value{
a matrix with correlations between traits
}
\description{
Function for computation of the correlation matrix from Cholesky factor
}
\examples{
NULL

}
