% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayes_helper.R
\name{gomp.a0}
\alias{gomp.a0}
\title{gomp.a0}
\usage{
gomp.a0(sampling = 1e+05, b_min = 0.02, b_max = 0.1, minimum_age = 15)
}
\arguments{
\item{sampling}{integer. Number of sampling steps. Default: 100000.}

\item{b_min}{numeric. Minimum of Gompertz \eqn{\beta} parameter. Default: 0.02.}

\item{b_max}{numeric. Maximum of Gompertz \eqn{\beta} parameter. Default: 0.1.}

\item{minimum_age}{numeric. Minimum age in years. Default: 15.}
}
\value{
A vector with coefficients for generating \eqn{\alpha} from
\eqn{\beta}, the parameters of the Gompertz function.
}
\description{
Internal function for generating starting values for the Gompertz model if
the starting age is not 15 years. Not run if the minimum age is actually 15.
The original formula derives from \emph{Sasaki and Kondo 2016}.
}
\references{
\insertRef{91048}{baytaAAR}
}
\keyword{internal}
