% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayes_helper.R
\name{hdi.agerange}
\alias{hdi.agerange}
\title{HDI median range of age estimates}
\usage{
hdi.agerange(x, age_identifier = "age.s")
}
\arguments{
\item{x}{a mcmc list.}

\item{age_identifier}{a character string of either "age.s" or "age.s_c" to
select the uncalibrated or calibrated age estimates. Default: "age.s".}
}
\value{
a vector with mean, median and mode of the HDI range
}
\description{
Comparison with known age-at-death
}
\examples{
NULL

}
