% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/known_age.R
\name{sequential.binom.test}
\alias{sequential.binom.test}
\title{Sequential output of cumulative binomial test}
\usage{
sequential.binom.test(
  mcmc_list,
  known_age,
  HDImass = 0.95,
  age_identifier = "age.s"
)
}
\arguments{
\item{mcmc_list}{MCMC output from coda chains.}

\item{known_age}{a vector of known age-at-death. NAs are allowed and those
entries will subsequently be ignored.}

\item{HDImass}{a numeric or a vector with the probability range.}

\item{age_identifier}{a character string of either "age.s" or "age.s_c" to
select the uncalibrated or calibrated age estimates. Default: "age.s".}
}
\value{
a data.frame
}
\description{
Sequential output of cumulative binomial test
}
\examples{
NULL

}
