% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{sorsum}
\alias{sorsum}
\title{Sorsum}
\format{
## `sorsum`
A data frame with 38 rows and 2 columns:
\itemize{
  \item \bold{id:} individual id
  \item \bold{auricular_surface:} auricular surface after Lovejoy et al. 1985
}
}
\description{
Neolithic population from a gallery grave in Hessen/Germany. From the
published raw data, the trait which was possible to assess most often
(auricular surface after Lovejoy et al. 1985) is included. It is taken from
Moser et al. 2025, table S69.
}
\references{
\insertRef{ref_128246}{baytaAAR}

\insertRef{ref_65832}{baytaAAR}
}
\seealso{
Other examples: 
\code{\link{spitalfields}}
}
\concept{examples}
