% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{spitalfields}
\alias{spitalfields}
\title{Spitalfields}
\format{
A data frame with 180 rows and 7 columns:
\itemize{
  \item \bold{Age:} Known age-at-death
  \item \bold{TO:} trait 'Transverse organization'
  \item \bold{ST:} trait 'Surface texture'
  \item \bold{MI:} trait 'Microporosity'
  \item \bold{MA:} trait 'Macroporosity'
  \item \bold{AP:} trait 'Apical changes'
  \item \bold{Sex:} Known sex
}
}
\source{
<http://faculty.las.illinois.edu/lylek/Oxford/Buckberry.csv>
}
\description{
Known age-at-death of an Early Modern crypt population. The data is also
available at the website of L. Konigsberg (s. link below).
}
\references{
\insertRef{ref_263525}{baytaAAR}
}
\seealso{
Other examples: 
\code{\link{sorsum_as}}
}
\concept{examples}
