% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayes_helper.R
\name{summed.prob.cat}
\alias{summed.prob.cat}
\title{Plots of stapled probability densities per category}
\usage{
summed.prob.cat(mcmc_list, age_identifier = "age.s", df_orig, group_col)
}
\arguments{
\item{mcmc_list}{a coda mcmc.list object.}

\item{age_identifier}{a character string of either "age.s" or "age.s_c" to
select the uncalibrated or calibrated age estimates. Default: "age.s".}

\item{df_orig}{a data.frame of the original raw data}

\item{group_col}{a string specifying the grouping category in `df_orig`.}
}
\value{
a ggplot object with stapled probability density plots
}
\description{
A plot of probability densities.
}
\examples{
NULL

}
