% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayes_helper.R
\name{threshold.chains}
\alias{threshold.chains}
\title{Compute thresholds for chains}
\usage{
threshold.chains(mcmc_list)
}
\arguments{
\item{mcmc_list}{MCMC output from coda chains.}
}
\value{
a mcmc-list of coda chains for threshold values of traits
}
\description{
Compute thresholds
}
\examples{
NULL

}
