% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayes_helper.R
\name{age.estim.summary}
\alias{age.estim.summary}
\title{Summary of age estimates}
\usage{
age.estim.summary(x, age_identifier = "age.s")
}
\arguments{
\item{x}{output from function \code{diagnostics.summary()}}

\item{age_identifier}{a character string of either "age.s" or "age.s_c" to
select the uncalibrated or calibrated age estimates. Default: "age.s".}
}
\value{
A data.frame with mean, median and mode as well as the HDI ranges as
specified in the output of \code{diagnostics.summary()} as columns and the
following rows:
\itemize{
\item{ \code{b} Gompertz parameter \eqn{\beta}. }
\item{ \code{a} Gompertz parameter \eqn{\alpha}. }
\item{ \code{M} Modal age, derived from the Gompertz parameters
   \eqn{\alpha} and \eqn{\beta} according to the equation
   (1 / \eqn{\beta}) * log(\eqn{\beta} / \eqn{\alpha}) + minimum_age. }
\item{ \code{age_mean} Mean ages.}
\item{ \code{hdi_diff} \emph{Highest density intervals}.}
}
}
\description{
Convenience function to quickly extract the age-related estimates from the
result of the function \code{diagnostics.summary()}.
}
\examples{
\dontshow{if (interactive()) withAutoprint(\{ # examplesIf}
# select Sorsum data with auricular surface after Lovejoy et al. 1985
sorsum <- sorsum_as[,2]

# example with default settings, please be a little bit patient
sorsum_res <- bay.ta(method = sorsum)

sorsum_diag <- diagnostic.summary(sorsum_res)

# show summary of age-related estimates
age.estim.summary(sorsum_diag)
\dontshow{\}) # examplesIf}
}
