% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bay_ta_jags.R
\name{bay.ta.jags}
\alias{bay.ta.jags}
\title{Bayesian Transition Analysis with JAGS}
\usage{
bay.ta.jags(
  method,
  parameters,
  gomp_b = NA,
  minimum_age = 15,
  maximum_age = 100,
  error_sd = NA,
  adaptSteps = 2000,
  burnInSteps = 3000,
  runjagsMethod = "rjags",
  nChains = 3,
  thinSteps = 1,
  numSavedSteps = 10000,
  silent.jags = F,
  silent.runjags = F,
  seed = seed
)
}
\arguments{
\item{method}{a matrix with traits, starting from 1.}

\item{parameters}{a vector with parameters to monitor.}

\item{gomp_b}{numeric. Optional Gompertz parameter. Default: NA.}

\item{minimum_age}{numeric. Minimum age. Default: 15.}

\item{maximum_age}{numeric. Maximum age. Default: 100.}

\item{error_sd}{numeric. Optional calibration term for estimation of age.
Default: NA.}

\item{adaptSteps}{integer. Number of adaptation steps. Default: 2000.}

\item{burnInSteps}{integer. Number of adaptation steps. Default: 3000.}

\item{runjagsMethod}{string. Mode to run runjags, options: "rjags",
"rjparallel", "parallel". Default: "rjags".}

\item{nChains}{integer. Number of chains. Default: 3.}

\item{thinSteps}{integer. Number of thinning steps. Default: 1.}

\item{numSavedSteps}{integer. Number of saved steps. Default: 10000.}

\item{silent.jags}{TRUE/FALSE Silent mode to run jags. Default: FALSE.}

\item{silent.runjags}{TRUE/FALSE Silent mode to run runjags. Default: FALSE.}
}
\value{
a coda object.

#@examples

result <- bay.ta.jags(method_matrix, c("beta", "age"))
}
\description{
This function implements a version of the Bayesian Transition Analysis with
MCMCM with JAGS. It presupposes an installed version of JAGS which it
interfaces with runjags (##).
}
