% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bay_ta_nimble.R
\name{bay.ta.nimble}
\alias{bay.ta.nimble}
\title{Bayesian Transition Analysis with NIMBLE}
\usage{
bay.ta.nimble(
  algorithm = "norm",
  cluster_n = 1,
  seed = FALSE,
  method,
  eta = 1,
  gomp_b = NA,
  minimum_age = 15,
  maximum_age = 100,
  parameters,
  nChains = 3,
  burnInSteps = 2000,
  thinSteps = 1,
  numSteps = 10000
)
}
\arguments{
\item{algorithm}{character string. Either \code{norm} for 'simple' ordered
regression or \code{mnorm} for multinormal ordered regression. Default:
\code{norm}.}

\item{cluster_n}{integer. Number of clusters for parallel processing chains,
maximally the number of cores -1. Default: 1.}

\item{seed}{integer. Random number for reproducibility. In parallel
processing, each cluster automatically get different seeds.}

\item{method}{matrix of integers. Ordinal trait(s) for age estimation.}

\item{eta}{double. Prior for the Cholesky factor, must be > 0. Only used for
multinormal ordered regression for the correlation matrix. 1 implies equal
correlations, lower values assume higher correlations. Default: 1.}

\item{gomp_b}{double. Optional prior for parameter Gompertz beta. Default:
NA.}

\item{minimum_age}{double. Minimum age for Gompertz distribution. Default:
15.}

\item{maximum_age}{double. Maximum age for Gompertz distribution. Default:
100.}

\item{parameters}{vector of character strings. Parameters to monitor.}

\item{nChains}{integer. Number of chains. Default: 4.}

\item{burnInSteps}{integer. Number of steps for burn-in. Default: 2000.}

\item{thinSteps}{integer. Thinning, i. e. which ith step should be saved.
Default: 1 (no thinning).}

\item{numSteps}{integer. Number of steps to run the model. Default: 10000.}
}
\value{
A coda object.
}
\description{
This is a wrapper for the functions \code{bay.ta.nimble.norm} and
\code{bay.ta.nimble.mnorm}. Both implement the Bayesian Transition Analysis with
MCMCM with the framework NIMBLE and allow the user to run models with
'simple' ordered regression or multinormal ordered regression, also with
parallel clusters.
}
\examples{
NULL


}
