% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bay_ta_nimble.R
\name{bay.ta.nimble}
\alias{bay.ta.nimble}
\title{Bayesian Transition Analysis with NIMBLE}
\usage{
bay.ta.nimble(
  algorithm,
  method,
  parameters,
  eta = 1,
  gomp_b = NA,
  error_sd = NA,
  minimum_age = 15,
  maximum_age = 100,
  burnInSteps = 2000,
  nChains = 3,
  thinSteps = 1,
  numSteps = 10000,
  seed = FALSE
)
}
\arguments{
\item{algorithm}{character string. Either \code{norm} for 'simple' ordered
regression or \code{mnorm} for multinormal ordered regression. Default:
\code{norm}.}

\item{method}{matrix of integers. Ordinal trait(s) for age estimation.}

\item{parameters}{vector of character strings. Parameters to monitor.}

\item{eta}{numeric. Parameter for the LKJ distribution, must be > 0. Only
used for multinormal ordered regression for the correlation matrix.
\code{1} implies equal correlations, lower values assume stronger
 correlations. Default: \code{1}.}

\item{gomp_b}{numeric. Optional prior for parameter Gompertz beta. Default:
\code{NA}.}

\item{error_sd}{numeric. Optional error parameter for age estimates. Default:
\code{NA}.}

\item{minimum_age}{numeric. Minimum age for Gompertz distribution. Default:
\code{15}.}

\item{maximum_age}{numeric. Maximum age for Gompertz distribution. Default:
\code{100}.}

\item{burnInSteps}{integer. Number of steps for burn-in. Default: \code{3000}.}

\item{nChains}{integer. Number of chains. Default: \code{3}.}

\item{thinSteps}{integer. Thinning, i. e. which \emph{i}th step should be
saved. Default: \code{1} (no thinning).}

\item{numSteps}{number of steps}

\item{seed}{integer. Random number for reproducibility. In parallel
processing, each cluster automatically gets different seeds. If no seed is
specified, the value is set to today's date as integer.}
}
\value{
A list of MCMC chains of class \code{coda::mcmc.list}.
}
\description{
Bayesian Transition Analysis with NIMBLE
}
\examples{
\dontshow{if (interactive()) withAutoprint(\{ # examplesIf}

  # select Sorsum data with auricular surface after Lovejoy et al. 1985 and
  # convert to matrix
  sorsum <- as.matrix(sorsum_as[,2])

  # example with default settings
  sorsum_res <- bay.ta(method = sorsum)
\dontshow{\}) # examplesIf}
}
\keyword{internal}
