% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bay_ta.R
\name{bay.ta}
\alias{bay.ta}
\title{Bayesian Transition Analysis with JAGS or NIMBLE}
\usage{
bay.ta(
  framework = "NIMBLE",
  algorithm = "norm",
  multicore = FALSE,
  seed = as.integer(format(Sys.Date(), "\%Y\%m\%d")),
  method,
  eta = 1,
  gomp_b = NA,
  error_sd = NA,
  minimum_age = 15,
  maximum_age = 100,
  parameters = c("b", "a", "beta0", "beta", "thresh", "age.s"),
  nChains = 3,
  adaptSteps = 2000,
  burnInSteps = 3000,
  thinSteps = 1,
  numSavedSteps = 10000,
  silent.jags = F,
  silent.runjags = F
)
}
\arguments{
\item{framework}{character string. Either \code{JAGS} or \code{NIMBLE}.
Default: \code{NIMBLE}.}

\item{algorithm}{character string. Either \code{norm} for 'simple' ordered
regression or \code{mnorm} for multinormal ordered regression. Default:
\code{norm}.}

\item{multicore}{\code{TRUE/FALSE}. If \code{TRUE} each chain is assigned to
a dedicated core. Default: \code{FALSE}.}

\item{seed}{integer. Random number for reproducibility. In parallel
processing, each cluster automatically gets different seeds. If no seed is
specified, the value is set to today's date as integer.}

\item{method}{matrix of integers, converted to matrix if not already matrix.
Ordinal trait(s) for age estimation.}

\item{eta}{numeric. Parameter for the LKJ distribution, must be > 0. Only
used for multinormal ordered regression for the correlation matrix.
\code{1} implies equal correlations, lower values assume stronger
 correlations. Default: \code{1}.}

\item{gomp_b}{numeric. Optional prior for parameter Gompertz beta. Default:
\code{NA}.}

\item{error_sd}{numeric. Optional error parameter for age estimates. Default:
\code{NA}.}

\item{minimum_age}{numeric. Minimum age for Gompertz distribution. Default:
\code{15}.}

\item{maximum_age}{numeric. Maximum age for Gompertz distribution. Default:
\code{100}.}

\item{parameters}{vector of character strings. Parameters to monitor.}

\item{nChains}{integer. Number of chains. Default: \code{3}.}

\item{adaptSteps}{integer. Number of adaptation steps, ignored when
\code{framework} is set to \code{NIMBLE}. Default: \code{2000}.}

\item{burnInSteps}{integer. Number of steps for burn-in. Default: \code{3000}.}

\item{thinSteps}{integer. Thinning, i. e. which \emph{i}th step should be
saved. Default: \code{1} (no thinning).}

\item{numSavedSteps}{integer. Number of saved steps. Default: \code{10000}.
The total number of steps equals \code{thinSteps × numSavedSteps}.}

\item{silent.jags}{TRUE/FALSE Silent mode to run JAGS. Default: \code{FALSE}.
Ignored when \code{framework} is set to \code{NIMBLE}.}

\item{silent.runjags}{TRUE/FALSE Silent mode to run runjags. Default:
\code{FALSE}. Ignored when \code{framework} is set to \code{NIMBLE}.}
}
\value{
A list of MCMC chains of class \code{coda::mcmc.list}.
}
\description{
\code{bay.ta()} implements latent trait analysis within a Bayesian Markov
Chain Monte Carlo (MCMC) framework. It is intended to estimate the
age-of-death of adult individuals for whom one or several ordinal traits
have been assessed. It produces probability densities for the individual
ages but also for the respective population as a whole. \code{bay.ta()} has
been introduced and tested in the paper ###.
\code{bay.ta()} is a wrapper for the functions \code{bay.ta.jags()} and
\code{bay.ta.nimble()}. NIMBLE allows the user to run models with multinormal
ordered regression, also with parallel clusters. In this respect, however,
JAGS tends to be more stable. The latter presupposes, however, that you have
installed JAGS outside of R.
}
\section{Data requirements}{

As input, \code{bay.ta()} assumes a \code{matrix} of trait expressions. In
its simplest form, this may contain only one column with a single trait. NAs
are allowed but neither must all entries in any of the rows be \code{NA} nor
can this be the case for one or several of the columns. \code{bay.ta} will
reject to run in such cases, and the offending rows or columns need to be
removed from analysis. Please see the article on Chelsea 'Old church' for an
example how this can be accomplished.
The levels of all traits must start at \code{1}. Binary traits are possible.
Mixing of levels like \code{1.5} as short-cut for a trait-expression between
\code{1} and \code{2}, however, should be an absolute no-go as this would
violate basic principles of ordinal scaling. Thus, for such cases a decision
for one of the neighboring levels has to be made or they need to be set to
\code{NA}.
The nodes (= rows of the matrix) do not have to be fully observed for the
multinormal model to run because with
\href{https://r-nimble.org/release-notes.html#february-14-2026-weve-released-version-1.4.1}{NIMBLE vers. 1.4.1.},
the NIMBLE team introduced a sampler for only partly observed multivariate
normal random variables.
}

\examples{
\dontshow{if (interactive()) withAutoprint(\{ # examplesIf}

  # select Sorsum data with auricular surface after Lovejoy et al. 1985
  sorsum <- sorsum_as[,2]

  # example with default settings
  sorsum_res <- bay.ta(method = sorsum)

  # example with framework JAGS
  sorsum_res <- bay.ta(framework = "JAGS", method = sorsum)

  # example with framework JAGS and multiple cores (parallel computing)
  sorsum_res <- bay.ta(framework = "JAGS", multicore = TRUE, method = sorsum)

  # example with 10,000 saved iterations and a thinning of 10 (= 100,000
  # iterations)
  sorsum_res <- bay.ta(method = sorsum, numSavedSteps = 10000, thin = 10)

  # select Spitalfields data with multiple traits
  spitalfields_traits <- spitalfields[,c(2:6)]

  # example with multinormal likelihood, please be patient
  spitalfields_res <- bay.ta(falgorithm = "mnorm",
  method = spitalfields_traits)
\dontshow{\}) # examplesIf}
}
