% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bay_ta.R, R/bay_ta_jags.R, R/bay_ta_nimble.R
\name{bay.ta}
\alias{bay.ta}
\alias{bay.ta.jags}
\alias{bay.ta.nimble}
\title{Bayesian Transition Analysis with JAGS or NIMBLE}
\usage{
bay.ta(
  framework = "NIMBLE",
  algorithm = "norm",
  multicore = FALSE,
  seed = as.integer(format(Sys.Date(), "\%Y\%m\%d")),
  method,
  eta = 1,
  gomp_b = NA,
  error_sd = NA,
  minimum_age = 15,
  maximum_age = 100,
  parameters = c("b", "a", "beta0", "beta", "thresh", "age.s"),
  nChains = 3L,
  adaptSteps = 2000,
  burnInSteps = 3000,
  thinSteps = 1,
  numSavedSteps = 10000,
  silent.jags = F,
  silent.runjags = F
)

bay.ta.jags(
  method,
  parameters,
  gomp_b = NA,
  minimum_age = 15,
  maximum_age = 100,
  error_sd = NA,
  adaptSteps = 2000,
  burnInSteps = 3000,
  runjagsMethod = "rjags",
  nChains = 3,
  thinSteps = 1,
  numSavedSteps = 10000,
  silent.jags = F,
  silent.runjags = F,
  seed = seed
)

bay.ta.nimble(
  algorithm,
  method,
  parameters,
  eta = 1,
  gomp_b = NA,
  error_sd = NA,
  minimum_age = 15,
  maximum_age = 100,
  burnInSteps = 2000,
  nChains = 3,
  thinSteps = 1,
  numSteps = 10000,
  seed = FALSE
)
}
\arguments{
\item{framework}{character string. Either \code{JAGS} \code{NIMBLE}. Default:
\code{NIMBLE}.}

\item{algorithm}{character string. Either \code{norm} for 'simple' ordered
regression or \code{mnorm} for multinormal ordered regression. Default:
\code{norm}.}

\item{multicore}{TRUE/FALSE. If TRUE each chain is assigned to a dedicated
core. Default: FALSE.}

\item{seed}{integer. Random number for reproducibility. In parallel
processing, each cluster automatically gets different seeds. If no seed is
specified, the value is set to today's date as integer.}

\item{method}{matrix of integers. Ordinal trait(s) for age estimation.}

\item{eta}{numeric. Prior for the Cholesky factor of the LKJ distribution,
must be > 0. Only used for multinormal ordered regression for the correlation
matrix. 1 implies equal correlations, lower values assume stronger
correlations. Default: 1.}

\item{gomp_b}{numeric. Optional prior for parameter Gompertz beta. Default:
NA.}

\item{error_sd}{numeric.}

\item{minimum_age}{numeric. Minimum age for Gompertz distribution. Default:
15.}

\item{maximum_age}{numeric. Maximum age for Gompertz distribution. Default:
100.}

\item{parameters}{vector of character strings. Parameters to monitor.}

\item{nChains}{integer. Number of chains. Default: 3.}

\item{adaptSteps}{integer. Number of adaptation steps, ignored when
\code{framework} is set to "NIMBLE". Default: 2000.}

\item{burnInSteps}{integer. Number of steps for burn-in. Default: 3000.}

\item{thinSteps}{integer. Thinning, i. e. which \emph{i}th step should be
saved. Default: 1 (no thinning).}

\item{numSavedSteps}{integer. Number of saved steps. Default: 10000. The
total number of steps equals \code{thinSteps × numSavedSteps}.}

\item{silent.jags}{TRUE/FALSE Silent mode to run JAGS. Default: FALSE.
Ignored when \code{framework} is set to "NIMBLE".}

\item{silent.runjags}{TRUE/FALSE Silent mode to run runjags. Default: FALSE.
Ignored when \code{framework} is set to "NIMBLE".}

\item{runjagsMethod}{string. Mode to run `runjags`, options: "rjags",
"rjparallel", "parallel". Default: "rjags".}

\item{numSteps}{number of steps}
}
\value{
A list of MCMC chains of class \code{coda::mcmc.list}
}
\description{
\code{bay.ta()} is a wrapper for the functions \code{bay.ta.jags()} and
\code{bay.ta.nimble()}. Both implement the Bayesian Transition Analysis with
MCMC. NIMBLE allows the user to run models with multinormal ordered regression,
also with parallel clusters. In this respect, however, JAGS tends to be more
stable. The latter presupposes, however, that you have installed JAGS outside
of R.

This function implements a version of the Bayesian Transition Analysis with
MCMCM with JAGS. It presupposes an installed version of JAGS which it
interfaces with `runjags` (##).
}
\details{
blabla
}
\examples{
\dontshow{if (interactive()) withAutoprint(\{ # examplesIf}

  # select Sorsum data with auricular surface after Lovejoy et al. 1985 and
  # convert to matrix
  sorsum <- as.matrix(sorsum_as[,2])

  # example with default settings
  sorsum_res <- bay.ta(method = sorsum)

  # example with framework JAGS
  sorsum_res <- bay.ta(framework = "JAGS", method = sorsum)

  # example with framework JAGS and multiple cores (parallel computing)
  sorsum_res <- bay.ta(framework = "JAGS", multicore = TRUE, method = sorsum)

  # example with 10,000 saved iterations and a thinning of 10 (= 100,000
  # iterations)
  sorsum_res <- bay.ta(method = sorsum, numSavedSteps = 10000, thin = 10)

  # select Spitalfields data with multiple traits and convert to matrix
  spitalfields_traits <- as.matrix(spitalfields[,c(2:6)])

  # example with multinormal likelihood
  spitalfields_res <- bay.ta(framework = "NIMBLE", algorithm = "mnorm",
  method = spitalfields_traits)
\dontshow{\}) # examplesIf}
}
