% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baytaAAR.R
\docType{package}
\name{baytaAAR-package}
\alias{baytaAAR}
\alias{baytaAAR-package}
\title{baytaAAR: Bayesian Transition Analysis with MCMC}
\description{
Provides Bayesian age estimation for bioarchaeological skeletal data using ordinal probit regression models implemented in JAGS and NIMBLE. The package is designed to handle multiple ordinal traits of adult individuals and incorporates a Gompertz prior on age to reflect population-level mortality trends. It accounts for observational error and supports full customization of model parameters and MCMC settings. Ideal for analyzing age-at-death in archaeological samples through a probabilistic and reproducible framework.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/ISAAKiel/baytaAAR}
  \item Report bugs at \url{https://github.com/ISAAKiel/baytaAAR/issues}
}

}
\keyword{internal}
