% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baytaAAR.R
\docType{package}
\name{baytaAAR-package}
\alias{baytaAAR}
\alias{baytaAAR-package}
\title{baytaAAR: Bayesian Transition Analysis with Markov Chain Monte Carlo}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Provides Bayesian age estimation for bioarchaeological skeletal data using ordinal probit regression models implemented in 'JAGS' and 'NIMBLE'. The package is designed to handle multiple ordinal traits of adult individuals and incorporates a Gompertz prior on age to reflect population-level mortality. It accounts for estimation uncertainties and supports full customization of model parameters and Markov Chain Monte Carlo settings.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/ISAAKiel/baytaAAR}
  \item \url{https://isaakiel.github.io/baytaAAR/}
  \item Report bugs at \url{https://github.com/ISAAKiel/baytaAAR/issues}
}

}
\author{
\strong{Maintainer}: Nils Müller-Scheeßel \email{nils.mueller-scheessel@ufg.uni-kiel.de} (\href{https://orcid.org/0000-0001-7992-8722}{ORCID}) [copyright holder]

Authors:
\itemize{
  \item Christoph Rinne \email{crinne@ufg.uni-kiel.de} (\href{https://orcid.org/0000-0002-9829-6182}{ORCID})
}

}
\keyword{internal}
