% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayes_helper.R
\name{diagnostic.summary}
\alias{diagnostic.summary}
\title{Diagnostic summary of MCMC samples}
\usage{
diagnostic.summary(mcmc_list, HDImass = 0.95, gelman_diag = TRUE)
}
\arguments{
\item{mcmc_list}{MCMC output from coda chains.}

\item{HDImass}{numeric. Value within 0 and 1. Default = 0.95.}

\item{gelman_diag}{logical. If TRUE, the gelman diagnostics for computing
the PSRF is invoked. Default: TRUE}
}
\value{
A data.frame of class `diagnostic_summary`.

Columns: PSRF Point est.,PSRF Upper C.I., Mean, Median, Mode, ESS, MCSE,
HDImass, HDIlow, HDIhigh.
}
\description{
Summarising diagnostics partly derived from
 \href{https://jkkweb.sitehost.iu.edu/DoingBayesianDataAnalysis/}{Kruschke
 (2015)}. Enter a codaMCMClist and receive a dataframe with summarising
 diagnostics.
}
\details{
Because the first threshold is always fixed, the gelman multivariate PSRF
 will always throw an error, this is automatically set to \code{FALSE}. If
 the gelman diagnostics still produce an error, deactivate \code{gelman_diag}
 altogether by setting it to \code{FALSE}, too.
}
\examples{
# select Sorsum data with auricular surface after Lovejoy et al. 1985 and
# convert to matrix
sorsum <- as.matrix(sorsum_as[,2])

# example with default settings
sorsum_res <- bay.ta(method = sorsum)

sorsum_diag <- diagnostic.summary(sorsum_res)

# show first rows
head(sorsum_diag)

}
