% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayes_helper.R
\name{diagnostic.summary}
\alias{diagnostic.summary}
\title{Diagnostic summary of MCMC samples}
\usage{
diagnostic.summary(
  codaMCMClist,
  HDImass = 0.95,
  gelman_diag = TRUE,
  gelman_diag_multivariate = FALSE
)
}
\arguments{
\item{HDImass}{numeric. Value within 0 and 1. Default = 0.95.}

\item{gelman_diag}{logical. If TRUE, the gelman diagnostics for computing
the PSRF is invoked. Default: TRUE}

\item{gelman_diag_multivariate}{logical. If TRUE, the gelman diagnostics for
computing the PSRF multivariately is invoked. Default: FALSE.}

\item{codaMCMClist.}{List in codaMCMClist format.}
}
\value{
A data.frame of class `diagnostic_summary`.
{Columns:} PSRF Point est.,PSRF Upper C.I., Mean, Median, Mode, ESS,
MCSE, HDImass, HDIlow, HDIhigh.
}
\description{
Summarising diagnostics partly derived from
 \href{https://jkkweb.sitehost.iu.edu/DoingBayesianDataAnalysis/}{Kruschke
 (2015)}. Enter a codaMCMClist and receive a dataframe with summarising
 diagnostics.
}
\details{
In certain situations, invoking the Gelman diagnostics can lead to errors,
 for example when constants are included or if the data is too sparse. In
 such cases, set first \code{gelman_diag_multivariate} to \code{FALSE}, and,
 if this still produces an error, deactivate \code{gelman_diag} altogether
 by setting it to \code{FALSE}, too.
}
\examples{
# run



}
