% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayes_helper.R
\name{prob.cat}
\alias{prob.cat}
\title{Summed or mean probability densities per category}
\usage{
prob.cat(
  mcmc_list,
  age_identifier = "age.s",
  df_orig,
  group_col,
  mode = c("mean", "summed")
)
}
\arguments{
\item{mcmc_list}{a coda mcmc.list object.}

\item{age_identifier}{a character string of either "age.s" or "age.s_c" to
select the uncalibrated or calibrated age estimates. Default: "age.s".}

\item{df_orig}{a data.frame of the original raw data}

\item{group_col}{a string specifying the grouping category in `df_orig`.}

\item{mode}{a string specifying the resulting data.frame of summed
probabilities or mean probabilities per category. Either `mean` or `summed`.}
}
\value{
a data.frame with either probability summed by category or mean
per category.
}
\description{
Summing or averaging probability densities per category.
}
\examples{
NULL

}
