% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/known_age.R
\name{sequential.binom.test}
\alias{sequential.binom.test}
\title{Sequential output of cumulative binomial test}
\usage{
sequential.binom.test(
  x,
  HDImass = 0.95,
  known_age,
  selector = NULL,
  age_identifier = "age.s"
)
}
\arguments{
\item{x}{a MCMC list}

\item{HDImass}{a numeric with the probability range.}

\item{known_age}{a vector with the known age-at-death.}

\item{selector}{a vector of individuals to include. Optional.}

\item{age_identifier}{a character string of either "age.s" or "age.s_c" to
select the uncalibrated or calibrated age estimates. Default: "age.s".}
}
\value{
a data.frame
}
\description{
Sequential output of cumulative binomial test
}
\examples{
NULL

}
