% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{spitalfields}
\alias{spitalfields}
\title{Spitalfields: Example dataset}
\format{
A data frame with 180 rows and 7 columns:
\itemize{
  \item \code{Age} known age-at-death
  \item \code{TO} trait 'Transverse organization'
  \item \code{ST} trait 'Surface texture'
  \item \code{MI} trait 'Microporosity'
  \item \code{MA} trait 'Macroporosity'
  \item \code{AP} trait 'Apical changes'
  \item \code{Sex} known sex
}
}
\source{
<http://faculty.las.illinois.edu/lylek/Oxford/Buckberry.csv>
}
\description{
Known age-at-death of an Early Modern crypt population. The data is also
available at the website of L. Konigsberg (s. link below).
}
\references{
\insertRef{ref_263525}{baytaAAR}
}
\seealso{
Other examples: 
\code{\link{sorsum_as}}
}
\concept{examples}
