% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayes_helper.R
\name{threshold.chains}
\alias{threshold.chains}
\title{Compute thresholds for chains}
\usage{
threshold.chains(x)
}
\arguments{
\item{x}{output from MCMC sampling}
}
\value{
a coda mcmc-list with three chains for threshold values of traits
}
\description{
Compute thresholds
}
\examples{
NULL

}
