% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayes_helper.R
\name{threshold.matrix}
\alias{threshold.matrix}
\title{Extract thresholds}
\usage{
threshold.matrix(x, mean_value = "Mode")
}
\arguments{
\item{x}{output from function `diagnostic.summary`}

\item{mean_value}{a character string of either "Mode", "Median" or "Mean".
Default: "Mode".}
}
\value{
a matrix with threshold values of traits
}
\description{
Extract thresholds
}
\examples{
NULL

}
